requirements.txt
google-adk==0.1.0
deprecated>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
joblib>=1.3.0
python-dotenv>=1.0.0
streamlit>=1.28.0
google-generativeai>=0.3.0
google-cloud-storage>=2.10.0 